fix BatchedParseBlock
